+++ packages/polaris-viz/src/components/Annotations/components/index.ts
@@ -1,0 +1,4 @@
+export {AnnotationLine} from './AnnotationLine';
+export {AnnotationLabel} from './AnnotationLabel';
+export {AnnotationContent} from './AnnotationContent';
+export {ShowMoreAnnotationsButton} from './ShowMoreAnnotationsButton';