+++ packages/polaris-viz/src/components/HorizontalBarChart/tests/HorizontalBarChart.test.tsx
@@ -1,12 +1,16 @@
 import React from 'react';
 import {mount} from '@shopify/react-testing';
 
-import {ChartContainer} from '../../ChartContainer';
 import {
   HorizontalBarChart,
   HorizontalBarChartProps,
 } from '../HorizontalBarChart';
 import {Chart} from '../Chart';
+
+jest.mock('@shopify/polaris-viz-core/src/utilities', () => ({
+  ...jest.requireActual('@shopify/polaris-viz-core/src/utilities'),
+  estimateStringWidth: jest.fn(() => 50),
+}));
 
 const mockProps: HorizontalBarChartProps = {
   data: [{name: 'Test', data: [{value: 10, key: 'data'}]}],